update user module

+++ src/app/routes/users/update/update.component.ts
@@ -1,11 +1,13 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute } from '@angular/router';
 import * as fromMenu from '@core/+state/selectors/menu.selectors';
 import * as fromProfile from '@core/+state/selectors/profile.selectors';
 import { Group, Language, Menu, Merchant, Profile } from '@models';
 import { Store } from '@ngrx/store';
-import { map, Observable, tap } from 'rxjs';
+import { ConfirmDialogComponent } from '@shared/components/confirm-dialog/confirm-dialog.component';
+import { map, Observable, take, tap } from 'rxjs';
 import * as UserActions from '../+state/actions/user.actions';
 import * as fromUser from '../+state/selectors/user.selectors';
 @Component({
@@ -23,21 +25,21 @@ export class UsersUpdateComponent implements OnInit {
   id$: Observable<number>;
   user$: Observable<Profile | undefined>;
 
-  constructor(private fb: FormBuilder, private store: Store, private route: ActivatedRoute) {
-    this.form = this.fb.nonNullable.group(
-      {
-        emailAddress: ['', [Validators.required, Validators.email]],
-        merchant: ['', [Validators.required]],
-        firstName: ['', [Validators.required]],
-        lastName: ['', [Validators.required]],
-        groups: [[], [Validators.required]],
-        defaultLanguage: ['en', [Validators.required]],
-        active: [true],
-        password: ['', [Validators.required]],
-        repeatPassword: ['', [Validators.required]],
-      },
-      { validators: [this.validatePassword] }
-    );
+  constructor(
+    private fb: FormBuilder,
+    private store: Store,
+    private route: ActivatedRoute,
+    public dialog: MatDialog
+  ) {
+    this.form = this.fb.nonNullable.group({
+      emailAddress: ['', [Validators.required, Validators.email]],
+      merchant: ['', [Validators.required]],
+      firstName: ['', [Validators.required]],
+      lastName: ['', [Validators.required]],
+      groups: [[], [Validators.required]],
+      defaultLanguage: ['en', [Validators.required]],
+      active: [true],
+    });
     this.menus$ = this.store.select(fromMenu.selectMenus);
     this.groups$ = this.store.select(fromProfile.selectGroups);
     this.languages$ = this.store.select(fromProfile.selectLanguages);
@@ -58,18 +60,6 @@ export class UsersUpdateComponent implements OnInit {
 
   ngOnInit() {}
 
-  validatePassword() {
-    return (form: FormGroup) => {
-      const password = form.get('password') as FormControl;
-      const repeatPassword = form.get('repeatPassword') as FormControl;
-      if (password.value !== repeatPassword.value) {
-        repeatPassword.setErrors({ notMatch: true });
-      } else {
-        repeatPassword.setErrors(null);
-      }
-    };
-  }
-
   compareGroup(a: Group, b: Group) {
     return a.id === b.id;
   }
@@ -79,6 +69,20 @@ export class UsersUpdateComponent implements OnInit {
   }
 
   delete(id: number) {
-    this.store.dispatch(UserActions.deleteUser({ id }));
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        title: 'Delete User',
+        message: 'Are you sure you want to delete this user?',
+        type: 'warning',
+      },
+    });
+    dialogRef
+      .afterClosed()
+      .pipe(take(1))
+      .subscribe(result => {
+        if (result) {
+          this.store.dispatch(UserActions.deleteUser({ id }));
+        }
+      });
   }
 }
